test(slack): align failure boundary coverage
