+++ packages/junior/tests/integration/oauth-resume-slack.test.ts
@@ -148,7 +148,7 @@ describe("oauth resume slack integration", () => {
       ),
     ).toEqual([
       expect.stringContaining(
-        "I ran into an internal error while processing that. Reference: `event_id=",
+        "I ran into an internal error while processing that.",
       ),
     ]);
   });
@@ -221,7 +221,7 @@ describe("oauth resume slack integration", () => {
     ).toEqual([
       "Connected. Continuing...",
       expect.stringContaining(
-        "I ran into an internal error while processing that. Reference: `event_id=",
+        "I ran into an internal error while processing that.",
       ),
     ]);
   });
@@ -371,42 +371,6 @@ describe("oauth resume slack integration", () => {
     expect(postCalls[1]?.params.text).toContain(
       getSlackInterruptionMarker().trim(),
     );
-    expect(postCalls[1]?.params.text).not.toContain("event_id=");
-  });
-
-  it("keeps the delivered resume reply when post-delivery commit fails", async () => {
-    const { resumeSlackTurn } = await import("@/chat/runtime/slack-resume");
-    const onFailure = vi.fn(async () => undefined);
-
-    await expect(
-      resumeSlackTurn({
-        messageText: "continue this turn",
-        conversationId: "slack:C123:1700000000.011",
-        turnId: "turn-resume-commit-fail",
-        channelId: "C123",
-        threadTs: "1700000000.011",
-        replyContext: {
-          credentialContext: {
-            actor: { type: "user", userId: "U123" },
-          },
-          destination: TEST_SLACK_DESTINATION,
-          source: testSlackSource("1700000000.011"),
-          actor: { platform: "slack", teamId: "T123", userId: "U123" },
-        },
-        agentRunner: modelReply("Final resumed answer"),
-        commitResult: async () => {
-          throw new Error("state write failed");
-        },
-        onFailure,
-      }),
-    ).rejects.toThrow("state write failed");
-
-    expect(onFailure).not.toHaveBeenCalled();
-    expect(
-      getCapturedSlackApiCalls("chat.postMessage").map(
-        (call) => call.params.text,
-      ),
-    ).toEqual([expect.stringContaining("Final resumed answer")]);
   });
 
   it("schedules plugin tasks after a successful resumed turn", async () => {
